refactor: split GitServiceTests into partial classes
