+++ GitTools.Tests/Services/GitServiceTagOperationsTests.cs
@@ -1,0 +1,119 @@
+using System.Diagnostics;
+using System.IO.Abstractions;
+using System.IO.Abstractions.TestingHelpers;
+using System.Reflection;
+using GitTools.Models;
+using GitTools.Services;
+using GitTools.Tests.Utils;
+using Spectre.Console;
+using Spectre.Console.Testing;
+
+namespace GitTools.Tests.Services;
+
+public sealed partial class GitServiceTests
+{
+
+    [Fact]
+    public async Task HasTagAsync_WhenTagExists_ShouldReturnTrue()
+    {
+        // Arrange
+        const string GIT_OUTPUT = "v1.0.0";
+
+        _fileSystem.Directory.Exists(Arg.Any<string>()).Returns(true);
+
+        _processRunner.RunAsync(Arg.Any<ProcessStartInfo>(), Arg.Any<DataReceivedEventHandler>(), Arg.Any<DataReceivedEventHandler>())
+            .Returns(static callInfo =>
+            {
+                var outputHandler = callInfo.ArgAt<DataReceivedEventHandler>(1);
+                outputHandler?.Invoke(null!, CreateDataReceivedEventArgs(GIT_OUTPUT));
+
+                return 0;
+            });
+
+        // Act
+        var result = await _gitService.HasTagAsync(REPO_PATH, TAG_NAME);
+
+        // Assert
+        result.ShouldBeTrue();
+
+        await _processRunner.Received(1).RunAsync
+        (
+            Arg.Is<ProcessStartInfo>(static psi =>
+                psi.FileName == "git" &&
+                psi.Arguments == $"tag -l {TAG_NAME}" &&
+                psi.WorkingDirectory == REPO_PATH),
+            Arg.Any<DataReceivedEventHandler>(),
+            Arg.Any<DataReceivedEventHandler>()
+        );
+    }
+
+    [Fact]
+    public async Task HasTagAsync_WhenTagDoesNotExist_ShouldReturnFalse()
+    {
+        // Arrange
+        _fileSystem.Directory.Exists(Arg.Any<string>()).Returns(true);
+
+        _processRunner.RunAsync(Arg.Any<ProcessStartInfo>(), Arg.Any<DataReceivedEventHandler>(), Arg.Any<DataReceivedEventHandler>())
+            .Returns(static callInfo =>
+            {
+                var outputHandler = callInfo.ArgAt<DataReceivedEventHandler>(1);
+                outputHandler?.Invoke(null!, CreateDataReceivedEventArgs(""));
+
+                return 0;
+            });
+
+        // Act
+        var result = await _gitService.HasTagAsync(REPO_PATH, TAG_NAME);
+
+        // Assert
+        result.ShouldBeFalse();
+    }
+
+    [Fact]
+    public async Task DeleteTagAsync_ShouldCallGitTagDelete()
+    {
+        // Arrange
+        _fileSystem.Directory.Exists(Arg.Any<string>()).Returns(true);
+
+        _processRunner.RunAsync(Arg.Any<ProcessStartInfo>(), Arg.Any<DataReceivedEventHandler>(), Arg.Any<DataReceivedEventHandler>())
+            .Returns(0);
+
+        // Act
+        await _gitService.DeleteTagAsync(REPO_PATH, TAG_NAME);
+
+        // Assert
+        await _processRunner.Received(1).RunAsync
+        (
+            Arg.Is<ProcessStartInfo>(static psi =>
+                psi.FileName == "git" &&
+                psi.Arguments == $"tag -d {TAG_NAME}" &&
+                psi.WorkingDirectory == REPO_PATH),
+            Arg.Any<DataReceivedEventHandler>(),
+            Arg.Any<DataReceivedEventHandler>()
+        );
+    }
+
+    [Fact]
+    public async Task DeleteRemoteTagAsync_ShouldCallGitPushOriginDelete()
+    {
+        // Arrange
+        _fileSystem.Directory.Exists(Arg.Any<string>()).Returns(true);
+
+        _processRunner.RunAsync(Arg.Any<ProcessStartInfo>(), Arg.Any<DataReceivedEventHandler>(), Arg.Any<DataReceivedEventHandler>())
+            .Returns(0);
+
+        // Act
+        await _gitService.DeleteRemoteTagAsync(REPO_PATH, TAG_NAME);
+
+        // Assert
+        await _processRunner.Received(1).RunAsync
+        (
+            Arg.Is<ProcessStartInfo>(static psi =>
+                psi.FileName == "git" &&
+                psi.Arguments == $"push origin :refs/tags/{TAG_NAME}" &&
+                psi.WorkingDirectory == REPO_PATH),
+            Arg.Any<DataReceivedEventHandler>(),
+            Arg.Any<DataReceivedEventHandler>()
+        );
+}
+}
